requirements.txt
praw
python-dotenv
tqdm
boto3
faust
